Test: Use isolatedModules: true (#108)
Changing ts-jest to `isolatedModules: true` makes our test suite
twice as fast. This is due to ts-jest not having to worry about
compiling TS at runtime when running tests.

This is a safe change to make because typescript tests are still
compiled normally with `yarn build`, in CI, and in vscode when
someone edits a test.